refactor(session-history): extract storage layer
Split the SQLite storage/schema/read-side helpers out of session_history.cpp
into a dedicated internal module so the facade file keeps the writer queue,
sampler, aggregators, and public API while the new storage unit owns DB open,
schema migration, CRUD, pruning, and row decoding.

Also wire the new files into the explicit CMake source list, refresh the
session-history unit tests to the current bitrate/delete APIs, narrow the
Windows/posix parent-directory ACL hardening so it only applies to the real
production session_history.db location, and fall back to the write connection
for read queries when a secondary read handle cannot be opened on this
machine.
